Verificando que petición sea de tipo ajax, para metodos storePhotos, deletePhotos, listPhotos

+++ app/Http/Controllers/ProductController.php
@@ -170,35 +170,44 @@ class ProductController extends Controller
         return view('products.edit-photo', compact('product'));
     }
 
-    public function listPhoto($productId)
+    public function listPhoto(Request $request, $productId)
     {
-        $product = $this->repository->get($productId);
-        if($product->hasPhotos())
+        if($request->ajax())
         {
-            $photos = $product->photos->toArray();
-            $this->setSuccess(true);
-            $this->addToResponseArray('photos', $photos);
+            $product = $this->repository->get($productId);
+            if($product->hasPhotos())
+            {
+                $photos = $product->photos->toArray();
+                $this->setSuccess(true);
+                $this->addToResponseArray('photos', $photos);
+            }
+            return $this->getResponseArrayJson();
         }
-        return $this->getResponseArrayJson();
     }
 
-    public function deletePhoto($photoId)
+    public function deletePhoto(Request $request, $photoId)
     {
-        $photo = $this->productPhotoRepository->get($photoId);
-        $this->productPhotoRepository->remove(
-            $photo->complete_path,
-            $photo->complete_thumbnail_path,
-            $photo->id
-        );
-        $this->setSuccess(true);
-        return $this->getResponseArrayJson();
+        if($request->ajax())
+        {
+           $photo = $this->productPhotoRepository->get($photoId);
+           $this->productPhotoRepository->remove(
+                $photo->complete_path,
+                $photo->complete_thumbnail_path,
+                $photo->id
+            );
+           $this->setSuccess(true);
+           return $this->getResponseArrayJson();
+       }
     }
 
     public function storePhoto(Request $request)
     {
-        $this->productPhotoRepository->register($request->file('file'), $request->input('productId'), 
-            \Auth::user()->id);
-        return response()->json(['status' => 'success', 'file' => $request->file('file')], 200);
+        if($request->ajax())
+        {
+            $this->productPhotoRepository->register($request->file('file'), $request->input('productId'), 
+                \Auth::user()->id);
+            return response()->json(['status' => 'success', 'file' => $request->file('file')], 200);  
+        }
     }
 
     public function dataTable()
